% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetCanada.R
\docType{data}
\name{datasetCanada}
\alias{datasetCanada}
\title{Temperature datasets from Canada.}
\format{
A list with two entries: \code{m_cood} and \code{m_data}.
\describe{
\item{\code{m_coord}}{a \code{tibble} with latitude, logitude and the name
of stations.}
\item{m_data}{a \code{tibble} where each column is the time series from a
weather station.}
}
}
\source{
the \code{CanadianWeather} dataset from the \code{R} package
\code{fda}.
}
\usage{
data("datasetCanada")
}
\description{
Temperature time series from 35 weather stations from Canada.  This dataset
is a classic one and was used in famous package \code{fda}. We have made a few
changes in this dataset.
}
\keyword{datasets}
