% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{fourier_b}
\alias{fourier_b}
\title{Smoothed curve in Fourier Series.}
\usage{
fourier_b(coef, x)
}
\arguments{
\item{coef}{Fourier coefficients.}

\item{x}{a time series to evaluate the smoothed curve.}
}
\value{
a time series with the smoothed curve.
}
\description{
This function computes the smoothed curve using Fourier coefficients.
}
\examples{
v_coef <- rnorm(23)
fourier_b(v_coef)
}
