% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikMultiNorm}
\alias{logLikMultiNorm}
\title{Log likelihood function for multivariate normal with spatial dependency.}
\arguments{
\item{mCoef}{coefficient matrix. Each column is the coefficient from a curve;}

\item{mDist}{distance matris;}

\item{s2}{variance from the covariance model;}

\item{phi}{variance from the covariance model;}

\item{rho}{variance from the covariance model;}
}
\description{
Log likelihood function for multivariate normal with spatial dependency.
}
